+++ server/.env
@@ -1,4 +1,4 @@
-DATABASE=mongodb+srv://aarib:mEeocdLHGM8GUAQu@donationplatformcluster.iu6tdc7.mongodb.net/?retryWrites=true&w=majority
+MONGO_URI="mongodb+srv://aarib:mEeocdLHGM8GUAQu@donationplatformcluster.iu6tdc7.mongodb.net/?retryWrites=true&w=majority"
 FRONTEND_URL=http://localhost:3000
 PORT=3001
 SECRET_KEY=MYNAMEISAARIB
