+++ server/db.js
@@ -1,0 +1,15 @@
+const mongoose = require("mongoose");
+
+module.exports = () => {
+  const connectionParams = {
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
+  };
+  try {
+    mongoose.connect(process.env.MONGO_URI, connectionParams);
+    console.log("Connected to database successfully");
+  } catch (error) {
+    console.log(error);
+    console.log("Could not connect database!");
+  }
+};
